move asteroid functions to separate file
