+++ lib/index.js
@@ -1,14 +1,14 @@
 const $ = require('jquery')
 const Ship = require('../lib/ship');
+const Asteroid = require('../lib/asteroids');
 
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
 var ship = new Ship();
-var largeAsteroidRadius = 20;
-var xWarpLeft = canvas.width + largeAsteroidRadius;
-var xWarpRight = -largeAsteroidRadius;
-var yWarpBottom = canvas.height + largeAsteroidRadius;
-var yWarpTop = -largeAsteroidRadius;
+
+function getRandomNumber(min, max){
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
 
 // create Asteroids
 var asteroids = [];
@@ -17,34 +17,6 @@ for (var x=0; x<5; x++){
   var newY = getRandomNumber(0, canvas.height)
   var newAsteroid = new Asteroid(newX, newY)
   asteroids.push(newAsteroid)
-};
-
-function getRandomNumber(min, max){
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-function Asteroid(x, y){
-  this.x = x;
-  this.y = y;
-  this.direction = getRandomNumber(0, 360)
-};
-
-Asteroid.prototype.draw = function () {
-  context.beginPath();
-  context.arc(this.x,this.y,largeAsteroidRadius,0,2*Math.PI);
-  context.stroke();
-  context.fillStyle = "gray";
-  context.fill();
-};
-
-Asteroid.prototype.move = function () {
-  this.x += Math.sin(this.direction);
-  if (this.x > xWarpLeft) this.x = xWarpRight;
-  if (this.x < xWarpRight) this.x = xWarpLeft;
-  
-  this.y -= Math.cos(this.direction);
-  if (this.y > yWarpBottom) this.y = yWarpTop;
-  if (this.y < yWarpTop) this.y = yWarpBottom;
 };
 
 var img = new Image();
@@ -58,9 +30,9 @@ requestAnimationFrame(function gameLoop() {
     asteroid.move();
     asteroid.draw();
   });
+  collisionDetection();
   requestAnimationFrame(gameLoop)
 });
-
 
 document.addEventListener('keydown', function (event) {
   console.log(ship.x, ship.y);
